requirements.txt
flask==2.0.1
google-generativeai>=0.1.0
selenium==4.0.0
webdriver-manager==3.8.6
lighthouse-parser==0.3.0
python-owasp-zap==0.1.0
pytest==6.2.5
allure-pytest==2.9.43